Trying to fix 32 bit int result on get for one machine

+++ src/plugins/macaddr/macaddr.c
@@ -45,7 +45,7 @@ void transformMac (Key * key)
 	}
 	macWithoutSeparators[12] = '\0';
 
-	long intValue = strtol (macWithoutSeparators, NULL, 16);
+	unsigned long long intValue = strtol (macWithoutSeparators, NULL, 16);
 
 	const int n = snprintf (NULL, 0, "%lu", intValue);
 	char * buffer = elektraMalloc (n + 1);
